+++ CHANGES.rst
@@ -6,6 +6,8 @@ Unreleased (2.9.0) (XXXX-XX-XX)
 
 - Remove layergroups from search.
 
+- Adapt search results to updated full-text search API response.
+
 
 Release 2.9.4 (2016-3-1)
 ---------------------